requirement.txt
torch
faster-whisper
fastapi[all]
moviepy
python-magic
torchvision
torchaudio
pydub
datasets
evaluate
librosa
soundfile
jiwer
transformers[torch]
tensorboardX
bitsandbytes
pytube
yt-dlp
huggingface_hub
lfs
langdetect
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib
aiofiles
gradio